backend/requirements.txt
fastapi>=0.111.0
uvicorn>=0.30.0
langchain>=0.2.0
langchain-google-genai>=1.0.5
langchain-community>=0.2.0
langgraph>=0.2.0
chromadb>=0.5.0
pymupdf>=1.24.0
tavily-python>=0.3.3
httpx>=0.27.0
python-dotenv>=1.0.1
pydantic>=2.7.0
python-multipart>=0.0.9
